consolidate interfaces
