+++ src/components/solution-contents/solution-contents.tsx
@@ -15,17 +15,8 @@
  */
 
 import { Component, Element, Event, EventEmitter, Host, h, Prop, VNode, Watch } from '@stencil/core';
-import { ISolutionItem } from '../solution-item/solution-item';
+import { IInventoryItem, ISolutionItem } from '../../utils/interfaces';
 import '@esri/calcite-components';
-
-export interface IInventoryItem {
-  id: string;
-  title: string;
-  dependencies?: IInventoryItem[];
-  type: string;
-  typeKeywords: string[];
-  solutionItem: ISolutionItem
-}
 
 @Component({
   tag: 'solution-contents',
