+++ src/components/solution-item-details/solution-item-details.tsx
@@ -18,17 +18,7 @@ import { Component, Element, h, Host, Listen, Prop, VNode } from '@stencil/core'
 import '@esri/calcite-components';
 import state from '../../utils/editStore';
 import { getProp } from '../../utils/common';
-
-export interface IItemDetails {
-  thumbnail: string;
-  title: string;
-  snippet: string;
-  description: string;
-  tags: string[];
-  accessInformation?: string;
-  licenseInfo?: string;
-  itemId: string
-}
+import { IItemDetails } from '../../utils/interfaces';
 
 @Component({
   tag: 'solution-item-details',
