+++ src/components/solution-spatial-ref/solution-spatial-ref.tsx
@@ -16,16 +16,10 @@
 
 import { Component, Element, h, Host, Method, Prop, State, Watch, VNode } from '@stencil/core';
 import '@esri/calcite-components';
-import { IWkidDescription, wkids } from './spatialreferences';
+import { wkids } from './spatialreferences';
 import state from '../../utils/editStore';
 import { getProp, nodeListToArray } from '../../utils/common';
-
-export interface ISpatialRefRepresentation {
-  display: string;
-  usingWkid: boolean;
-  wkid: number;
-  wkt: string;
-}
+import { ISpatialRefRepresentation, IWkidDescription } from '../../utils/interfaces';
 
 @Component({
   tag: 'solution-spatial-ref',
